+++ CHANGELOG.md
@@ -6,10 +6,9 @@ Dragonboat v3.0 is a major release with breaking changes. Please make sure to ca
 
 - Added on disk state machine (statemachine.IOnDiskStateMachine) support. The on disk state machine is close to the concept described in the section 5.2 of Diego Ongaro's Raft thesis. 
 - Added new API for requesting a  snapshot to be created or exported.
-- Added the ability to use exported snapshot to repair permanently damaged cluster that no longer has its majority quorum.
+- Added the ability to use exported snapshot to repair permanently damaged cluster that no longer has majority quorum.
 - Added new API for cleaning up data and release disk spaces after a node is removed from its Raft cluster.
 - Added the ability to limit peak memory usage when disk or network is slow.
-- Added Go module support. Go 1.12 is required.
 
 ### Improvements